Create 4. leaky ReLU
It seems like the main problem was in the huge size of the layers compared to the batch size.
So the program couldn't generalize and just remembered how to multiply the sample batch.
When I made the batch size=200 and number of neurons per layer = 20, it became much smarter in terms of generalization.
Batch size 7 and number of neurons 100 didn't work well.😄
So today I used leaky_relu, though it did'nt help.😄
I wanted to make a program with big amount of layers, but I tried to fix it, and the layers have gone.😄
Maybe I'll create big ass layers NN later. Wanna learn dropout technique though.