Create 7. great usual program y=2x (imporving with SAM)
A program that calculates y=2x (like in the 6th program here), but it uses SAM (I mostly wrote it, looking at the ChatGPT example, need some time to undertand the logic fully). The use of sharpness aware minimization (at least in the form I used) doesnt make sense in this task, so the program is not very good.